refactor: move mapeo commands to new module, preparing for rpc bridge
